feat: 상품 아이템 위젯을 추가하고, 판매 정보 관리 뷰모델 및 글쓰기 유효성 검증 로직을 구현했습니다.

+++ lib/core/utils/validation/write_validation.dart
@@ -31,7 +31,7 @@ class WriteValidation {
     return null;
   }
 
-  static String? validatePrice(double? value) {
+  static String? validatePrice(int? value) {
     if (value == null) {
       return '가격을 입력해 주세요. (0원 가능)';
     }
